server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.15
PyYAML==6.0.1
